+++ src/layouts/AppShell/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+export type AppShellPropsType = {
+  children: ReactNode;
+  header?: ReactNode;
+  footer?: ReactNode;
+  bottomNavigation?: ReactNode;
+  extras?: ReactNode;
+  withNotification?: boolean;
+  className?: string;
+};