+++ src/layouts/index.ts
@@ -1,0 +1,4 @@
+export * from "./AppShell";
+export * from "./AuthShell";
+export * from "./DashboardFooter";
+export * from "./DashboardHeader";