+++ src/views/FeatureUnavailableView/index.ts
@@ -1,0 +1,2 @@
+export { FeatureUnavailableView } from "./FeatureUnavailableView";
+export type { FeatureUnavailableViewPropsType } from "./types";