+++ src/views/NotFoundView/index.ts
@@ -1,0 +1,2 @@
+export { NotFoundView } from "./NotFoundView";
+export type { NotFoundViewPropsType } from "./types";