+++ src/views/index.ts
@@ -1,0 +1,2 @@
+export * from "./NotFoundView";
+export * from "./FeatureUnavailableView";